optimize test setup

+++ src/main/java/com/hotpads/datarouter/test/node/basic/sorted/test/SortedNodeIntegrationTests.java
@@ -54,9 +54,9 @@ public class SortedNodeIntegrationTests{
 	@Parameters
 	public static Collection<Object[]> parameters(){
 		List<Object[]> params = ListTool.create();
-//		params.add(new Object[]{DRTestConstants.CLIENT_drTestHibernate0, HibernateClientType.INSTANCE, false});
-//		params.add(new Object[]{DRTestConstants.CLIENT_drTestJdbc0, JdbcClientType.INSTANCE, false});
-//		params.add(new Object[]{DRTestConstants.CLIENT_drTestHBase, HBaseClientType.INSTANCE, false});
+		params.add(new Object[]{DRTestConstants.CLIENT_drTestHibernate0, HibernateClientType.INSTANCE, false});
+		params.add(new Object[]{DRTestConstants.CLIENT_drTestJdbc0, JdbcClientType.INSTANCE, false});
+		params.add(new Object[]{DRTestConstants.CLIENT_drTestHBase, HBaseClientType.INSTANCE, false});
 		params.add(new Object[]{DRTestConstants.CLIENT_drTestHBase, HBaseClientType.INSTANCE, true});
 		return params;
 	}
@@ -66,11 +66,13 @@ public class SortedNodeIntegrationTests{
 		this.router = new SortedBasicNodeTestRouter(clientName, getClass(), entity);
 		this.node = router.sortedBeanSorted();
 		this.entityNode = router.sortedBeanEntity();
-		resetTable();
 	}
 	
 
-	public void resetTable(){
+	public void resetTable(boolean force){
+		long numExistingDatabeans = IterableTool.count(node.scan(null, null));
+		if(!force && TOTAL_RECORDS == numExistingDatabeans){ return; }
+		
 		node.deleteAll(null);
 		List<SortedBean> remainingAfterDelete = ListTool.createArrayList(node.scan(null, null));
 		Assert.assertEquals(0, CollectionTool.size(remainingAfterDelete));
@@ -165,6 +167,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetEntity(){
+		resetTable(false);
 		if(!isHBaseEntity()){ return; }
 		SortedBeanEntityKey ek1 = new SortedBeanEntityKey(S_albatross, S_ostrich);
 		SortedBeanEntity albatrossOstrich = entityNode.getEntity(ek1, null);
@@ -176,12 +179,14 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetAll(){
+		resetTable(false);
 		List<SortedBean> allBeans = ListTool.createArrayList(node.scan(null, null));
 		Assert.assertEquals(TOTAL_RECORDS, CollectionTool.size(allBeans));
 	}
 	
 	@Test
 	public synchronized void testGetFirstKey(){
+		resetTable(false);
 		SortedBeanKey firstKey = node.getFirstKey(null);
 		Assert.assertEquals(STRINGS.first(), firstKey.getA());
 		Assert.assertEquals(STRINGS.first(), firstKey.getB());
@@ -191,6 +196,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetFirst(){
+		resetTable(false);
 		SortedBean firstBean = node.getFirst(null);
 		Assert.assertEquals(STRINGS.first(), firstBean.getKey().getA());
 		Assert.assertEquals(STRINGS.first(), firstBean.getKey().getB());
@@ -200,6 +206,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetWithPrefix(){
+		resetTable(false);
 		//first 3 fields fixed
 		SortedBeanKey prefix1 = new SortedBeanKey(STRINGS.first(), STRINGS.last(), 2, null);
 		List<SortedBean> result1 = node.getWithPrefix(prefix1, false, null);
@@ -222,6 +229,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetWithPrefixes(){
+		resetTable(false);
 		SortedBeanKey prefixA = new SortedBeanKey(STRINGS.first(), PREFIX_a, null, null);
 		SortedBeanKey prefixCh = new SortedBeanKey(STRINGS.first(), PREFIX_ch, null, null);
 		List<SortedBeanKey> prefixes = ListTool.create(prefixA, prefixCh);
@@ -235,6 +243,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetKeysInRange(){
+		resetTable(false);
 		SortedBeanKey alp1 = new SortedBeanKey(RANGE_alp, null, null, null);
 		SortedBeanKey emu1 = new SortedBeanKey(RANGE_emu, null, null, null);
 		List<SortedBeanKey> result1 = node.getKeysInRange(alp1, true, emu1, true, null);
@@ -257,6 +266,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetInRange(){
+		resetTable(false);
 		SortedBeanKey alp1 = new SortedBeanKey(RANGE_alp, null, null, null);
 		SortedBeanKey emu1 = new SortedBeanKey(RANGE_emu, null, null, null);
 		List<SortedBean> result1 = node.getRange(alp1, true, emu1, true, null);
@@ -279,6 +289,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test //small batch sizes to make sure we're resuming each batch from the correct spot
 	public synchronized void testIncrementalScan(){
+		resetTable(false);
 		Config smallIterateBatchSize = new Config().setIterateBatchSize(3);
 
 		int expectedSize1 = RANGE_LENGTH_alp_emu_inc * NUM_ELEMENTS * NUM_ELEMENTS * NUM_ELEMENTS;
@@ -310,6 +321,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testPrefixedRange(){
+		resetTable(false);
 		if(isHBaseEntity()){ return; }//not implemented
 		SortedBeanKey prefix = new SortedBeanKey(PREFIX_a, null, null, null);
 		SortedBeanKey al = new SortedBeanKey(RANGE_al, null, null, null);
@@ -321,6 +333,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGet(){
+		resetTable(false);
 		Iterable<SortedBean> iterable = node.scan(null, null);
 		for(SortedBean sortedBeanFromScan : iterable){
 			SortedBean sortedBeanFromGet = node.get(sortedBeanFromScan.getKey(), null);
@@ -330,6 +343,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testGetMulti(){
+		resetTable(false);
 		Iterable<SortedBean> iterable = node.scan(null, null);
 		Set<SortedBean> allBeans = Sets.newHashSet(iterable);
 		Assert.assertEquals(TOTAL_RECORDS, allBeans.size());
@@ -342,6 +356,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testFullScanKeys(){
+		resetTable(false);
 		Iterable<SortedBeanKey> iterable = node.scanKeys(null, null);
 		long numKeys = IterableTool.count(iterable);
 		Assert.assertEquals(TOTAL_RECORDS, numKeys);
@@ -349,6 +364,7 @@ public class SortedNodeIntegrationTests{
 	
 	@Test
 	public synchronized void testFullScan(){
+		resetTable(false);
 		Iterable<SortedBean> iterable = node.scan(null, null);
 		long numDatabeans = IterableTool.count(iterable);
 		Assert.assertEquals(TOTAL_RECORDS, numDatabeans);
@@ -356,6 +372,7 @@ public class SortedNodeIntegrationTests{
 
 	@Test
 	public synchronized void testDelete(){
+		resetTable(true);
 		int remainingElements = TOTAL_RECORDS;
 		
 		//delete
@@ -383,7 +400,7 @@ public class SortedNodeIntegrationTests{
 		remainingElements -= NUM_PREFIX_a * NUM_ELEMENTS * NUM_ELEMENTS * NUM_ELEMENTS;
 		Assert.assertEquals(remainingElements, IterableTool.count(node.scan(null, null)).intValue());
 
-		resetTable();//in case this one doesn't run last
+		resetTable(false);//leave the table full
 	}
